change file structure to accomodate new overview layout
