+++ src/components/Routes/Fund/FundOverview/FundOverview.tsx
@@ -24,7 +24,6 @@ export const FundOverview: React.FC<FundOverviewProps> = ({ address }) => (
     <GridRow>
       <GridCol>
         <FundPerformanceTable address={address} />
-        <FundMetrics address={address} />
       </GridCol>
     </GridRow>
     <GridRow>
